backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
groq>=0.9.0
httpx>=0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.1